Add files via upload
A program that allows users to store and manage contact information. The program provides options to add a new contact by entering their name, phone number, and email address. It also allow users to view their contact list, edit existing contacts, and delete contacts if needed. The program can also Store the contacts in memory or in a file for persistent storage.